fix: Kritische Bugfixes + Import-Support für attribute_views/price_types/relation_types + Doku
Bugfixes:
- Doppelte Transaktions-Verschachtelung in JsonFormatImporter entfernt
- filesize() TypeError bei strict_types=1 behoben (int-Cast)
- Autorisierungsprüfung in ExportJobController für show/update/destroy/execute/download
- try/catch für Import-Fehlerbehandlung in JsonExportImportController
- Excel-Export loggt Warnung wenn Filter gesetzt aber nicht unterstützt

Import-Erweiterung:
- Neue Import-Handler: importAttributeViews, importPriceTypes, importRelationTypes
- Sheet-Definitionen in SheetParser (15_Attribut_Sichten, 16_Preistypen, 17_Beziehungstypen)
- JSON-zu-Sheet-Mapping in JsonFormatImporter für alle 3 Sektionen
- Excel-Export-Template erweitert um 3 neue Sheets

Dokumentation:
- 12-json-export-import.md: Vollständige JSON-Schema-Dokumentation (de/en)
- HelpView: Neue Sektionen für JSON Export/Import und Export-Jobs

+++ app/Console/Commands/JsonExport.php
@@ -84,7 +84,7 @@ class JsonExport extends Command
         $exporter->exportToFile($outputPath, $sections, $filters);
 
         $duration = round(microtime(true) - $startTime, 2);
-        $size = $this->formatBytes(filesize($outputPath));
+        $size = $this->formatBytes((int) filesize($outputPath));
 
         $this->newLine();
         $this->info("Export abgeschlossen!");
